requirements.txt
fastapi>=0.115.8
pydantic>=2.11.3
pydantic-settings>=2.8.1
python-dotenv>=1.0.1
redis>=5.2.0
fastapi-cache2>=0.2.2
slowapi>=0.1.6
python-jose>=3.3.0
passlib>=1.7.4
supabase>=2.10.0
google-generativeai>=0.8.4
reportlab>=3.6.12
PyMuPDF>=1.25.3
docx2txt>=0.8
fastapi-limiter>=0.1.6
langchain>=0.3.19
langchain-google-genai>=2.0.9
langchain-community>=0.3.18
FAISS-cpu>=1.10.0
python-pptx>=1.0.2
json-repair>=0.7.0
openai>=1.3.0
python-multipart>=0.0.20
uvicorn>=0.34.0
email-validator>=2.2.0
httpx>=0.24.1
google-genai>=1.13.0
aiofiles>=24.1.0